Prepare to convert various types to size_t.
This makes some behaviorally-invariant changes to make certain code that
currently only works correctly with signed types work safely regardless of the
signedness of the types in question.  This is preparation for a future change
that will convert a variety of types to size_t.

There are also some formatting changes (e.g. converting "enum hack" usage to real consts) to make it simpler to just change "int" to "size_t" in the future to change the types of those constants.

BUG=none
R=andrew@webrtc.org, juberti@webrtc.org, kwiberg@webrtc.org
TBR=ajm

Review URL: https://codereview.webrtc.org/1174813003

Cr-Commit-Position: refs/heads/master@{#9413}

+++ webrtc/common_audio/resampler/sinusoidal_linear_chirp_source.h
@@ -40,7 +40,7 @@ class SinusoidalLinearChirpSource : public SincResamplerCallback {
     kMinFrequency = 5
   };
 
-  double sample_rate_;
+  int sample_rate_;
   int total_samples_;
   double max_frequency_;
   double k_;
